Move error to top, create MappingsTable component, move everything into Mappings folder
